do not show errors before user interaction
Errors aren't shown before a user interaction. If user interacts with a field
errors for that field are shown. If user submits the form, errors for all field
of the form are shown.

What is treated as an user interaction differs by field. Only if an interaction
could be seen as finished, it's causing errors to be shown. All focus out events
are treated as finished user interactions. A change event on an input field isn't
to prevent errors from popping up, while user inserts a valid string (e.g. if a
minimum length is required). On the other hand an change on select and checkbox
is.

This topic was discussed intensively in #15. There were many different ideas how
it could be implemented. This PR tries to reflect the last state of the debate.

In opposite to branch feature/delayed-validations this is not optional. I don't
think there are use cases where you want to show errors for untouched fields to
the user. If there are use cases one could pass `true` to `showErrors` property
of fields.
There shouldn't be any changes for people using server-side validations, since
there these ones aren't present before a form is submitted.