[FEATURE] Adding onKeyUp action
### Why?

We should be exposing a way for users to hook into the `keyup` event

### How does this resolve the issue?

This PR adds support for an `onKeyUp` action. It’s simply passed up
from the widget, into the field, and potentially above. I’ve also added
a simple test for this and have bumped the version numbers.